+++ crates/hwpforge-smithy-hwp5/src/layout_hint_patch.rs
@@ -195,6 +195,7 @@ fn collect_flow_paragraph_layout_hints(
             | Hwp5Control::Curve(_)
             | Hwp5Control::ConnectLine(_)
             | Hwp5Control::Equation(_)
+            | Hwp5Control::Dutmal(_)
             | Hwp5Control::OleObject(_)
             | Hwp5Control::Unknown { .. } => {}
         }
@@ -229,6 +230,7 @@ fn collect_scope_paragraph_layout_hints(paragraph: &Hwp5Paragraph, scope: &mut S
             | Hwp5Control::Curve(_)
             | Hwp5Control::ConnectLine(_)
             | Hwp5Control::Equation(_)
+            | Hwp5Control::Dutmal(_)
             | Hwp5Control::OleObject(_)
             | Hwp5Control::Unknown { .. } => {}
         }